+++ src/client/interfaces/Markup.ts
@@ -31,15 +31,23 @@ export declare namespace InlineKeyboardButton {
     /** If set, pressing the button will prompt the user to select one of their chats, open that chat and insert the bot's username and the specified inline query in the input field. May be empty, in which case just the bot's username will be inserted. Not supported for messages sent in channel direct messages chats and on behalf of a Telegram Business account. */
     switch_inline_query: string;
   }
+<<<<<<< HEAD
   export interface SwitchInlineCurrentChatButton
     extends AbstractInlineKeyboardButton {
+=======
+  export interface SwitchInlineCurrentChatButton extends AbstractInlineKeyboardButton {
+>>>>>>> v4
     /** If set, pressing the button will insert the bot's username and the specified inline query in the current chat's input field. May be empty, in which case only the bot's username will be inserted.
 
     This offers a quick way for the user to open your bot in inline mode in the same chat - good for selecting something from multiple options. Not supported in channels and for messages sent in channel direct messages chats and on behalf of a Telegram Business account. */
     switch_inline_query_current_chat: string;
   }
+<<<<<<< HEAD
   export interface SwitchInlineChosenChatButton
     extends AbstractInlineKeyboardButton {
+=======
+  export interface SwitchInlineChosenChatButton extends AbstractInlineKeyboardButton {
+>>>>>>> v4
     /** If set, pressing the button will prompt the user to select one of their chats of the specified type, open that chat and insert the bot's username and the specified inline query in the input field. Not supported for messages sent in channel direct messages chats and on behalf of a Telegram Business account. */
     switch_inline_query_chosen_chat: SwitchInlineQueryChosenChat;
   }