+++ src/structures/chat/BackgroundFill.js
@@ -54,11 +54,19 @@ class BackgroundFill {
   isGradient() {
     return Boolean(
       "topColor" in this &&
+<<<<<<< HEAD
         this.topColor &&
         "bottomColor" in this &&
         this.bottomColor &&
         "rotationAngle" in this &&
         this.rotationAngle,
+=======
+      this.topColor &&
+      "bottomColor" in this &&
+      this.bottomColor &&
+      "rotationAngle" in this &&
+      this.rotationAngle,
+>>>>>>> v4
     );
   }